move lifecycle callbacks type definitions into their respective middleware
